+++ src/main/java/chess/ChessGame.java
@@ -5,9 +5,9 @@ import chess.domain.ScoreCalculator;
 import chess.domain.CommandType;
 import chess.domain.GameCommand;
 import chess.domain.piece.Color;
-import chess.domain.piece.NormalPiecesGenerator;
+import chess.domain.piece.generator.NormalPiecesGenerator;
 import chess.domain.piece.Piece;
-import chess.domain.piece.PiecesGenerator;
+import chess.domain.piece.generator.PiecesGenerator;
 import chess.view.InputView;
 import chess.view.ResultView;
 import java.util.List;
