+++ src/main/java/chess/domain/piece/generator/PiecesGenerator.java
@@ -1,5 +1,6 @@
-package chess.domain.piece;
+package chess.domain.piece.generator;
 
+import chess.domain.piece.Piece;
 import chess.domain.position.Position;
 import java.util.Map;
