+++ src/test/java/chess/domain/ChessBoardTest.java
@@ -9,11 +9,11 @@ import chess.domain.piece.Color;
 import chess.domain.piece.EmptyPiece;
 import chess.domain.piece.King;
 import chess.domain.piece.Knight;
-import chess.domain.piece.NormalPiecesGenerator;
+import chess.domain.piece.generator.NormalPiecesGenerator;
 import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceName;
-import chess.domain.piece.PiecesGenerator;
+import chess.domain.piece.generator.PiecesGenerator;
 import chess.domain.piece.Queen;
 import chess.domain.piece.Rook;
 import chess.domain.position.Column;
